Update & rework dependencies
