+++ sim/pyproject.toml
@@ -8,7 +8,7 @@ version = "0.0.0"
 description = "Example simulations and benchmarks for blop (not published to PyPI)"
 readme = "README.md"
 authors = [{ name = "Thomas Hopkins", email = "thopkins1@bnl.gov" }]
-requires-python = ">=3.10"
+requires-python = ">=3.11"
 dependencies = [
   "blop",
   "ophyd-async",
