+++ src/blop/__init__.py
@@ -1,4 +1,3 @@
-from .ax import DOF, Agent, ChoiceDOF, DOFConstraint, Objective, OutcomeConstraint, RangeDOF, ScalarizedObjective
 from .plans import acquire_baseline, default_acquire, optimize, optimize_step, sample_suggestions
 
 try:
@@ -8,14 +7,6 @@ except ImportError:
 
 __all__ = [
     "__version__",
-    "Agent",
-    "ChoiceDOF",
-    "DOF",
-    "DOFConstraint",
-    "Objective",
-    "OutcomeConstraint",
-    "RangeDOF",
-    "ScalarizedObjective",
     "acquire_baseline",
     "default_acquire",
     "optimize",
