+++ src/blop/ax/__init__.py
@@ -1,13 +1,13 @@
-from ..queueserver import OptimizationResult
-from .agent import Agent as Agent
-from .agent import QueueserverAgent as QueueserverAgent
-from .dof import DOF, ChoiceDOF, DOFConstraint, RangeDOF
-from .objective import Objective, OutcomeConstraint, ScalarizedObjective, to_ax_objective_str
-from .optimizer import AxOptimizer
+try:
+    from .agent import Agent as Agent
+    from .dof import DOF, ChoiceDOF, DOFConstraint, RangeDOF
+    from .objective import Objective, OutcomeConstraint, ScalarizedObjective, to_ax_objective_str
+    from .optimizer import AxOptimizer
+except ImportError as e:
+    raise ImportError("The ax integration requires additional dependencies. Install them with: pip install blop[ax]") from e
 
 __all__ = [
     "Agent",
-    "QueueserverAgent",
     "DOF",
     "RangeDOF",
     "ChoiceDOF",
@@ -17,5 +17,4 @@ __all__ = [
     "ScalarizedObjective",
     "to_ax_objective_str",
     "AxOptimizer",
-    "OptimizationResult",
 ]
