+++ src/blop/bayesian/__init__.py
@@ -1,0 +1,10 @@
+try:
+    from .kernels import LatentKernel
+    from .models import LatentGP
+except ImportError as e:
+    raise ImportError("The bayesian module requires additional dependencies. Install them with: pip install blop[ax]") from e
+
+__all__ = [
+    "LatentKernel",
+    "LatentGP",
+]
